Move parameter configurations
