+++ .mypy-strict.ini
@@ -99,3 +99,6 @@ ignore_missing_imports = True
 
 [mypy-deprecation.*]
 ignore_missing_imports = True
+
+[mypy-ert.config.*]
+ignore_errors = True
