+++ .mypy.ini
@@ -132,3 +132,7 @@ ignore_missing_imports = True
 
 [mypy-ert.callbacks]
 ignore_errors = True
+
+[mypy-ert.config.*]
+ignore_missing_imports = True
+ignore_errors = True
