+++ src/ert/_c_wrappers/enkf/enkf_main.py
@@ -23,13 +23,13 @@ from numpy.random import SeedSequence
 
 from ert._c_wrappers.analysis.configuration import UpdateConfiguration
 from ert._c_wrappers.enkf.analysis_config import AnalysisConfig
-from ert._c_wrappers.enkf.config.parameter_config import ParameterConfig
 from ert._c_wrappers.enkf.enkf_obs import EnkfObs
 from ert._c_wrappers.enkf.ensemble_config import EnsembleConfig
 from ert._c_wrappers.enkf.enums import RealizationStateEnum
 from ert._c_wrappers.enkf.ert_run_context import RunContext
 from ert._c_wrappers.enkf.model_config import ModelConfig
 from ert._c_wrappers.util.substitution_list import SubstitutionList
+from ert.config.parameter_config import ParameterConfig
 from ert.job_queue import WorkflowRunner
 from ert.runpaths import Runpaths
