+++ src/ert/callbacks.py
@@ -4,9 +4,10 @@ import logging
 from pathlib import Path
 from typing import Callable, Iterable, Tuple
 
-from ert._c_wrappers.enkf import EnsembleConfig, RunArg, SummaryConfig
-from ert._c_wrappers.enkf.config.parameter_config import ParameterConfig
+from ert._c_wrappers.enkf import EnsembleConfig, RunArg
 from ert._c_wrappers.enkf.enums import RealizationStateEnum
+from ert.config import SummaryConfig
+from ert.config.parameter_config import ParameterConfig
 
 from .load_status import LoadResult, LoadStatus
