+++ src/ert/logging/logger.conf
@@ -112,7 +112,7 @@ loggers:
   ert.analysis:
     level: DEBUG
     propagate: yes
-  ert._c_wrappers.enkf.config:
+  ert.configig:
     level: DEBUG
     propagate: yes
   ert._c_wrappers.enkf.enkf_main:
