+++ tests/unit_tests/c_wrappers/res/enkf/test_ensemble_config.py
@@ -7,7 +7,7 @@ import xtgeo
 from ecl.summary import EclSum
 
 from ert._c_wrappers.enkf import ConfigKeys, EnsembleConfig, ErtConfig
-from ert._c_wrappers.enkf.config.gen_data_config import GenDataConfig
+from ert.config.gen_data_config import GenDataConfig
 from ert.parsing import ConfigValidationError, ConfigWarning
